requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cleantext==1.1.4
click==8.1.3
colorama==0.4.6
Django==4.1.13
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.10.7
gunicorn==20.1.0
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.3
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
nltk==3.8.1
numpy==1.24.2
outscraper==3.0.3
packaging==23.0
pandas==1.5.3
Pillow==10.2.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.4
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.18
Wand==0.6.11
whitenoise==6.4.0
yarl==1.9.1
